+++ changelog.md
@@ -1,6 +1,6 @@
 # graphql-react changelog
 
-## Next
+## 18.0.0
 
 ### Major